backend/requirements.txt
Flask==2.3.3
streamlit==1.30.0
requests==2.31.0
langchain==0.1.16
langchain-google-genai==0.0.9
google-generativeai>=0.3.1,<0.4.0
python-dotenv==1.0.0
